removed the new directories and replaced the detection files with the new ones
