--- extras/threshold.cpp
@@ -1,52 +1,0 @@
-#include <iostream>
-#include <cv.h>
-#include <highgui.h>
-#include <ros/ros.h>
-#include "std_msgs/String.h"
-#include "std_msgs/Int8.h"
-#include <fstream>
-#include <vector>
-#include <std_msgs/Bool.h>
-#include <opencv2/core/core.hpp>
-#include <opencv2/highgui/highgui.hpp>
-#include <opencv2/opencv.hpp>
-#include "std_msgs/Float32MultiArray.h"
-#include <opencv/highgui.h>
-#include <image_transport/image_transport.h>
-#include "std_msgs/Float64MultiArray.h"
-#include <cv_bridge/cv_bridge.h>
-#include <sstream>
-#include <string>
-
-using namespace std;
-using namespace cv;
-
-int main(int argc,char **argv){
-
-		cv::inRange(frame,hsv_min,hsv_max,dst);
-		///////then apply the filter////////////
-		std::vectors<std::vectors<cv::Points>> contours;
-		findContours(dst, contours, CV_RETR_TREE, CV_CHAIN_APPROX_SIMPLE);  // Find the contours
-    	
-    	for (int i = 0; i < contours.size(); i++)  // iterate through each contour.
-      {
-        double a = contourArea(contours[i], false);  //  Find the area of contour
-        if (a > largest_area)
-        {
-          largest_area = a;
-          largest_contour_index = i;  // Store the index of largest contour
-        }
-      }
-      // Convex HULL
-      std::vector<std::vector<cv::Point> > hull(contours.size());
-      convexHull(cv::Mat(contours[largest_contour_index]), hull[largest_contour_index], false);
-
-      std::vector<cv::Point2f> center(1);
-      std::vector<float> radius(1);
-      cv::minEnclosingCircle(contours[largest_contour_index], center[0], radius[0]);
-      cv::Point2f pt;
-      pt.x = 320;  // size of my screen
-      pt.y = 240;
-
-
-}
