+++ task_handler_layer/task_buoy/cfg/buoy.cfg
@@ -5,11 +5,11 @@ from dynamic_reconfigure.parameter_generator_catkin import *
 
 gen = ParameterGenerator()
 
-gen.add("t1min_param",    int_t,    0, "t1min parameter", 126,  0, 260)
-gen.add("t1max_param",    int_t,    0, "t1max parameter", 200,  0, 260)
+gen.add("t1min_param",    int_t,    0, "t1min parameter", 0,  0, 260)
+gen.add("t1max_param",    int_t,    0, "t1max parameter", 0,  0, 260)
 gen.add("t2min_param",    int_t,    0, "t2min parameter", 0,  0, 260)
-gen.add("t2max_param",    int_t,    0, "t2max parameter", 255,  0, 260)
-gen.add("t3min_param",    int_t,    0, "t3min parameter", 0,  0, 260)
+gen.add("t2max_param",    int_t,    0, "t2max parameter", 0,  0, 260)
+gen.add("t3min_param",    int_t,    0, "t3min parameter", 100,  0, 260)
 gen.add("t3max_param",    int_t,    0, "t3max parameter", 255,  0, 260)
 
 exit(gen.generate(PACKAGE, "task_buoy", "buoy"))
