+++ task_handler_layer/task_gate/cfg/gate.cfg
@@ -6,9 +6,9 @@ from dynamic_reconfigure.parameter_generator_catkin import *
 gen = ParameterGenerator()
 
 gen.add("t1min_param",    int_t,    0, "t1min parameter", 0,  0, 260)
-gen.add("t1max_param",    int_t,    0, "t1max parameter", 130,  0, 260)
+gen.add("t1max_param",    int_t,    0, "t1max parameter", 100,  0, 260)
 gen.add("t2min_param",    int_t,    0, "t2min parameter", 0,  0, 260)
-gen.add("t2max_param",    int_t,    0, "t2max parameter", 100,  0, 260)
-gen.add("t3min_param",    int_t,    0, "t3min parameter", 150,  0, 260)
+gen.add("t2max_param",    int_t,    0, "t2max parameter", 50,  0, 260)
+gen.add("t3min_param",    int_t,    0, "t3min parameter", 80,  0, 260)
 gen.add("t3max_param",    int_t,    0, "t3max parameter", 260,  0, 260)
 exit(gen.generate(PACKAGE, "task_gate", "gate"))
