+++ task_handler_layer/task_line/cfg/line.cfg
@@ -7,11 +7,11 @@ gen = ParameterGenerator()
 
 gen.add("orange_param",    int_t,    0,
         "Orange percent parameter", 7,  0, 100)
-gen.add("t1min_param",    int_t,    0, "t1min parameter", 121,  0, 260)
-gen.add("t1max_param",    int_t,    0, "t1max parameter", 198,  0, 260)
+gen.add("t1min_param",    int_t,    0, "t1min parameter", 0,  0, 260)
+gen.add("t1max_param",    int_t,    0, "t1max parameter", 80,  0, 260)
 gen.add("t2min_param",    int_t,    0, "t2min parameter", 0,  0, 260)
-gen.add("t2max_param",    int_t,    0, "t2max parameter", 255,  0, 260)
-gen.add("t3min_param",    int_t,    0, "t3min parameter", 0,  0, 260)
-gen.add("t3max_param",    int_t,    0, "t3max parameter", 255,  0, 260)
+gen.add("t2max_param",    int_t,    0, "t2max parameter", 260,  0, 260)
+gen.add("t3min_param",    int_t,    0, "t3min parameter", 20,  0, 260)
+gen.add("t3max_param",    int_t,    0, "t3max parameter", 260,  0, 260)
 
 exit(gen.generate(PACKAGE, "task_line", "line"))
